Extract CreateResponder.
