+++ app/controllers/vocabularies_controller.rb
@@ -1,10 +1,7 @@
-require 'json/ld'
-
 class VocabulariesController < ApplicationController
   before_filter :load_vocab, :only => :show
 
   def index
-
   end
 
   def new
@@ -13,16 +10,5 @@ class VocabulariesController < ApplicationController
 
   def create
     VocabularyCreator.call(params[:vocabulary], CreateResponder.new(self))
-  end
-
-  class CreateResponder < SimpleDelegator
-    def success(vocabulary)
-      redirect_to term_path(vocabulary)
-    end
-
-    def failure(vocabulary)
-      __getobj__.instance_variable_set(:@vocabulary, vocabulary)
-      render :new
-    end
   end
 end
